requirements.txt
nonebot2[fastapi,websockets,ujson]
nonebot-adapter-onebot
jmcomic
Pillow
img2pdf
